+++ plansys2_executor/src/plansys2_executor/Utils.cpp
@@ -157,7 +157,7 @@ std::shared_ptr<DurativeAction> get_action_from_string(
 
     for (size_t i = 0; i < action_output->parameters.size(); i++) {
       std::string pattern = "?" + std::to_string(i);
-      int pos;
+      size_t pos;
       while ((pos = at_start_req.find(pattern)) != std::string::npos) {
         at_start_req.replace(pos, pattern.length(), action_output->parameters[i].name);
       }
@@ -185,7 +185,7 @@ std::shared_ptr<DurativeAction> get_action_from_string(
 
     for (size_t i = 0; i < action_output->parameters.size(); i++) {
       std::string pattern = "?" + std::to_string(i);
-      int pos;
+      size_t pos;
       while ((pos = at_start_req.find(pattern)) != std::string::npos) {
         at_start_req.replace(pos, pattern.length(), action_output->parameters[i].name);
       }
